+++ devito/ir/iet/__init__.py
@@ -4,4 +4,3 @@ from devito.ir.iet.visitors import *  # noqa
 from devito.ir.iet.utils import *  # noqa
 from devito.ir.iet.analysis import *  # noqa
 from devito.ir.iet.scheduler import *  # noqa
-from devito.ir.iet.builtins import *  # noqa
